Add Hackerland Radio Tranmissters solution
Hackerland is a one-dimensional city with n  houses,
where each house i is located at some x(i) on the x-axis.
The Mayor wants to install radio transmitters on the roofs
of the city's houses. Each transmitter has a range, k,
meaning it can transmit a signal to all houses <= k units
of distance away.

Given a map of Hackerland and the value of k, can you find
and print the minimum number of transmitters needed to cover
every house in the city? (Every house must be covered by at
least one transmitter) Each transmitter must be installed on
top of an existing house.