Determining DNA health
DNA is a nucleic acid present in the bodies of living things. Each
piece of DNA contains a number of genes, some of which are beneficial
and increase the DNA's total health. Each gene has a health value,
and the total health of a DNA is the sum of the health values of all
the beneficial genes that occur as a substring in the DNA. We represent
genes and DNA as non-empty strings of lowercase English alphabetic
letters, and the same gene may appear multiple times as a susbtring of a DNA.

https://www.hackerrank.com/challenges/determining-dna-health/problem