Abbreviation - partially solved
You can perform the following operations on the string, a:

Capitalize zero or more of a's lowercase letters.
Delete all of the remaining lowercase letters in a.

Given two strings,  a and b, determine if it's possible to make
a equal to b as described. If so, print YES on a new line. Otherwise,
print NO.

https://www.hackerrank.com/challenges/abbr/problem